service/service-user/scripts/requirements.txt
python-docx>=1.1.2
pdf2docx>=0.5.8
